Proxy the dependencies in the URL generation command
With Magento 2.1.8+ CLI commands are constructed when the CLI client is
invoked. Additionally, DI will construct any objects that those commands
type in their constructors, so on and so fourth down the stack.

The problem is, this extension may exist in an environment in which the
store is not yet installed, and the database is thus unavailable. A
common place this is used is in 2.1.x in CI, as the database is required
in order to generate static content.

This commit proxies the dependencies of this constructor, such that the
constructor can be used (and thus appear in the CLI menu) but not
invoked while the store is not installed.