Allow running the command once again
Recently, Magneto was modified such that it throws an exception if the
area is modified during application runtime, if it has been set
previously. This extension sets it prior to regenerating the URLs.

Additional work attempted to repair this change by checking if the area
code was set. However, this has not functioned as expected (at least in
the most recent version of Magento), as Magento throws an exception on
querying an area code that has not already been set.

This commit modifies the setter such that it attempts to set the
application state during runtime, and on failure, preserves it as is.
This allows upstream to make the decision as to what state should be
set, but provides a sane fallback in the case it does not.